+++ operator.ps1
@@ -153,6 +153,7 @@ if ( $IsWindows ){
 }
 . $pwshSrcDir/Decrease-Indent_function.ps1
 . $pwshSrcDir/Unzip-Archive_function.ps1
+. $pwshSrcDir/Get-ClipboardAlternative_function.ps1
 
 ## file watcher
 . $pwshSrcDir/fwatch_function.ps1